stocky-web/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
pydantic==2.9.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
aiosqlite==0.20.0
groq==0.25.0
kiteconnect==5.0.1
pyotp==2.9.0
requests==2.32.3
python-dotenv==1.2.1
yfinance==1.1.0
feedparser==6.0.12
nsetools==2.0.1
numpy==2.2.0
